requirements.txt
google-adk>=1.2.0
google-cloud-build>=3.0.0
requests>=2.31.0
flask>=3.0.0
gunicorn>=22.0.0
